Split out a couple more util functions, absolute import paths
